requirements.txt
streamlit==1.32.0
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.2.1
openpyxl==3.1.2
google-generativeai>=0.4.1
python-dotenv==1.0.1
langchain==0.1.16
langchain-core==0.1.53
langchain-google-genai==0.0.11
tabulate==0.9.0
xlsxwriter==3.1.9
numpy==1.26.4
xlsxwriter
langchain-community
protobuf
chromadb==0.4.24
json5==0.9.14
demjson3==3.0.6